Formatting

+++ SteeringNode/main.c
@@ -44,25 +44,25 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 FUSES = {.low = 0xFF, .high = 0xD9, .extended = 0xFD};
 #endif
 
-#define ARR_LEN(arr)	(sizeof(arr) / sizeof(arr[0]))
+#define ARR_LEN(arr)    (sizeof(arr) / sizeof(arr[0]))
 
-#define SHIFT_LIGHT_PORT	PORTE
-#define SHIFT_LIGHT_R		PIN4 // Red rgb light
-#define SHIFT_LIGHT_B		PIN3 // Blue rgb light
+#define SHIFT_LIGHT_PORT    PORTE
+#define SHIFT_LIGHT_R       PIN4 // Red rgb light
+#define SHIFT_LIGHT_B       PIN3 // Blue rgb light
 
-#define RPM_PORT		PORTB
-#define RPM_PIN			PIN4
-#define RPM_TCCR		TCCR2A
-#define RPM_OCR			OCR2A
-#define RPM_WGM1		WGM21
-#define RPM_WGM0		WGM20
-#define RPM_CS1			CS21
-#define RPM_CS0			CS20
-#define RPM_COMA1		COM0A1
-#define RPM_COMA0		COM0A0
+#define RPM_PORT        PORTB
+#define RPM_PIN         PIN4
+#define RPM_TCCR        TCCR2A
+#define RPM_OCR         OCR2A
+#define RPM_WGM1        WGM21
+#define RPM_WGM0        WGM20
+#define RPM_CS1         CS21
+#define RPM_CS0         CS20
+#define RPM_COMA1       COM0A1
+#define RPM_COMA0       COM0A0
 
-#define RPM_MAX_VALUE	13000
-#define RPM_MIN_VALUE	3300
+#define RPM_MAX_VALUE   13000
+#define RPM_MIN_VALUE   3300
 
 
 static void rx_complete(uint8_t mob);
@@ -70,8 +70,8 @@ static void tx_complete(uint8_t mob);
 static void can_default(uint8_t mob);
 
 static int32_t map(int32_t x,
-				   const int32_t from_low, const int32_t from_high,
-				   const int32_t to_low, const int32_t to_high) {
+                   const int32_t from_low, const int32_t from_high,
+                   const int32_t to_low, const int32_t to_high) {
 	if (x < from_low) {
 		x = from_low;
 	} else if (x > from_high) {
@@ -87,7 +87,7 @@ static void set_rpm(int16_t rpm) {
 	// too soon. The value is determined by increasing the calibration value
 	// until it "looked right".
 	const int8_t calibration = 80;
-	RPM_OCR = map(rpm, RPM_MIN_VALUE, RPM_MAX_VALUE, 0, 0xFF-calibration);
+	RPM_OCR = map(rpm, RPM_MIN_VALUE, RPM_MAX_VALUE, 0, 0xFF - calibration);
 }
 
 int main(void) {
@@ -109,9 +109,9 @@ int main(void) {
 
 	// init Timer0 PWM PB4 for the RPM-counter
 	{
-		RPM_TCCR |= (1<<RPM_WGM1)|(1<<RPM_WGM0); 	// Fast PWM mode
-		RPM_TCCR |= (1<<RPM_CS1)|(1<<RPM_CS1); 		// F_CPU/64 prescalar
-		RPM_TCCR |= (1<<RPM_COMA1)|(0<<RPM_COMA0);	// Clear RPM_OCR on compare match. Set RPM_OCR at TOP.
+		RPM_TCCR |= (1 << RPM_WGM1) | (1 << RPM_WGM0); // Fast PWM mode
+		RPM_TCCR |= (1 << RPM_CS1) | (1 << RPM_CS1); // F_CPU/64 prescalar
+		RPM_TCCR |= (1 << RPM_COMA1) | (0 << RPM_COMA0); // Clear RPM_OCR on compare match. Set RPM_OCR at TOP.
 
 		set_rpm(0);
 		SET_PIN_MODE(RPM_PORT, RPM_PIN, OUTPUT);
@@ -127,9 +127,9 @@ int main(void) {
 
 	usart1_printf("\n\n\nSTARTING\n");
 
-	sei();										//Enable interrupt
+	sei();                                      //Enable interrupt
 
-	while(1) {
+	while (1) {
 		// Main work loop
 
 		// First lets store the current status of the paddleshifters
@@ -150,17 +150,17 @@ int main(void) {
 		{
 			// Test the raw diplay
 			for (int num = 0; num <= 9; ++num) {
-				for (int digit = 0; digit < 7; ++digit){
+				for (int digit = 0; digit < 7; ++digit) {
 					const char ascii_num = num + '0'; // Raw number to ascii
-					seg7_disp_char(digit, ascii_num, (num%2 == 0) ?
-						true : false);
+					seg7_disp_char(digit, ascii_num, (num % 2 == 0) ?
+					               true : false);
 				}
-					_delay_ms(125);
+				_delay_ms(125);
 			}
 
 			// Test the high level display string with dot
-			for (int i = 0; i < 150; ++i){
-				char buff[3+1+1] = {'\0'}; // 3 digits 1 dot and 1 \0
+			for (int i = 0; i < 150; ++i) {
+				char buff[3 + 1 + 1] = {'\0'}; // 3 digits 1 dot and 1 \0
 				snprintf(buff, ARR_LEN(buff), "%d.", i);
 				seg7_disp_str(buff, 0, 2);
 				_delay_ms(25);
@@ -182,25 +182,25 @@ int main(void) {
 				DMUX_Y6,
 				DMUX_Y7,
 			};
-			for (int i = 0; i < ARR_LEN(leds); ++i) {
+			for (int i = 0; i < (int)ARR_LEN(leds); ++i) {
 				// When setting one of these low we allow current to flow from
 				// the LED to GND thus the LED will light up.
 				// In short LOW == ON
 
 				dmux_set_y_low(leds[i]);
-				seg7_disp_char(3, i+'0', false);
+				seg7_disp_char(3, i + '0', false);
 
 				{
-					set_rgb_color(RED); 		_delay_ms(50);
-					set_rgb_color(GREEN); 		_delay_ms(50);
-					set_rgb_color(BLUE); 		_delay_ms(50);
+					set_rgb_color(RED);         _delay_ms(50);
+					set_rgb_color(GREEN);       _delay_ms(50);
+					set_rgb_color(BLUE);        _delay_ms(50);
 
-					set_rgb_color(YELLOW); 		_delay_ms(50);
-					set_rgb_color(MAGENTA); 	_delay_ms(50);
-					set_rgb_color(CYAN); 		_delay_ms(50);
+					set_rgb_color(YELLOW);      _delay_ms(50);
+					set_rgb_color(MAGENTA);     _delay_ms(50);
+					set_rgb_color(CYAN);        _delay_ms(50);
 
-					set_rgb_color(WHITE); 		_delay_ms(50);
-					set_rgb_color(COLOR_OFF); 	_delay_ms(50);
+					set_rgb_color(WHITE);       _delay_ms(50);
+					set_rgb_color(COLOR_OFF);   _delay_ms(50);
 				}
 			}
 		}
@@ -214,7 +214,7 @@ int main(void) {
 			for (int16_t rpm = 0; rpm <= RPM_MAX_VALUE; rpm += 100) {
 				set_rpm(rpm);
 
-				snprintf(buff, ARR_LEN(buff), "%d.", rpm/100);
+				snprintf(buff, ARR_LEN(buff), "%d.", rpm / 100);
 				seg7_disp_str(buff, 4, 6);
 				_delay_ms(25);
 			}
@@ -223,7 +223,7 @@ int main(void) {
 			for (int16_t rpm = RPM_MAX_VALUE; rpm >= 0; rpm -= 100) {
 				set_rpm(rpm);
 
-				snprintf(buff, ARR_LEN(buff), "%d.", rpm/100);
+				snprintf(buff, ARR_LEN(buff), "%d.", rpm / 100);
 				seg7_disp_str(buff, 4, 6);
 				_delay_ms(25);
 			}
@@ -231,7 +231,7 @@ int main(void) {
 #endif
 	}
 
-    return 0;
+	return 0;
 }
 
 static void rx_complete(uint8_t mob) {
@@ -253,11 +253,11 @@ static void rx_complete(uint8_t mob) {
 }
 
 static void tx_complete(uint8_t mob) {
-	MOB_ABORT();					// Freed the MOB
-	MOB_CLEAR_INT_STATUS();			// and reset MOb status
-	CAN_DISABLE_MOB_INTERRUPT(mob);	// Unset interrupt
+	MOB_ABORT();                    // Freed the MOB
+	MOB_CLEAR_INT_STATUS();         // and reset MOb status
+	CAN_DISABLE_MOB_INTERRUPT(mob); // Unset interrupt
 }
 
 static void can_default(uint8_t mob) {
-	MOB_CLEAR_INT_STATUS(); 		// and reset MOb status
+	MOB_CLEAR_INT_STATUS();         // and reset MOb status
 }
